requirements.txt
streamlit
python-dotenv

langchain
langchain-community
langchain-text-splitters

langgraph

faiss-cpu

sentence-transformers

groq

reportlab

pypdf

tiktoken